pyproject.toml
[tool.poetry]
name = "anime_fastapi"
version = "0.1.0"
description = ""
authors = ["UserCommon <kirilnenashev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
python-dotenv = "^0.20.0"
uvicorn = "^0.18.2"
asyncpg = "^0.25.0"
SQLAlchemy = "^1.4.39"
fastapi-users = {extras = ["sqlalchemy"], version = "^10.1.1"}
sqladmin = "^0.1.11"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
main = "anime_fastapi.main:main"